Add swagger in drf, fixed endpoints related at accounts, fernet key has been moved in env file, create custom authentication

+++ password_manager/users/views.py
@@ -1,3 +1,4 @@
+from drf_spectacular.utils import extend_schema, OpenApiResponse, OpenApiExample
 from rest_framework import status, permissions
 from rest_framework.authtoken.models import Token
 from rest_framework.response import Response
@@ -9,6 +10,49 @@ from users.serializers import CustomUserSerializer
 class RegisterView(APIView):
     permission_classes = [permissions.AllowAny]
 
+    @extend_schema(
+        request={
+            'application/json': {
+                'schema': {
+                    'type': 'object',
+                    'properties': {
+                        'username': {'type': 'string'},
+                        'password': {'type': 'string'},
+                    },
+                    'required': ['username', 'password'],
+                }
+            }
+        },
+        responses={
+            201: OpenApiResponse(
+                description='Successful register',
+                response={
+                    'type': 'object',
+                    'properties': {
+                        'token': {'type': 'string'},
+                    },
+                },
+            ),
+            400: OpenApiResponse(
+                description='Bad request',
+                response={
+                    'type': 'object',
+                    'properties': {
+                        'error': {'type': 'string'},
+                    },
+                },
+            ),
+        },
+        examples=[
+            OpenApiExample(
+                'Example request',
+                value={
+                    'username': 'your_username',
+                    'password': 'your_password',
+                },
+            ),
+        ],
+    )
     def post(self, request):
         serializer = CustomUserSerializer(data=request.data)
         if serializer.is_valid():
@@ -21,6 +65,49 @@ class RegisterView(APIView):
 class LoginView(APIView):
     permission_classes = [permissions.AllowAny]
 
+    @extend_schema(
+        request={
+            'application/json': {
+                'schema': {
+                    'type': 'object',
+                    'properties': {
+                        'username': {'type': 'string'},
+                        'password': {'type': 'string'},
+                    },
+                    'required': ['username', 'password'],
+                }
+            }
+        },
+        responses={
+            200: OpenApiResponse(
+                description='Successful login',
+                response={
+                    'type': 'object',
+                    'properties': {
+                        'token': {'type': 'string'},
+                    },
+                },
+            ),
+            400: OpenApiResponse(
+                description='Bad request',
+                response={
+                    'type': 'object',
+                    'properties': {
+                        'error': {'type': 'string'},
+                    },
+                },
+            ),
+        },
+        examples=[
+            OpenApiExample(
+                'Example request',
+                value={
+                    'username': 'your_username',
+                    'password': 'your_password',
+                },
+            ),
+        ],
+    )
     def post(self, request):
         username = request.data.get('username')
         password = request.data.get('password')
@@ -34,6 +121,33 @@ class LoginView(APIView):
 class LogoutView(APIView):
     permission_classes = [permissions.IsAuthenticated]
 
+    @extend_schema(
+        summary="Logout user",
+        description="Logs out the currently authenticated user by deleting their authentication token.",
+        responses={
+            200: OpenApiResponse(
+                description="Successfully logged out.",
+                response={
+                    "type": "object",
+                    "properties": {
+                        "message": {"type": "string", "example": "Successfully logged out."},
+                    },
+                },
+            ),
+            401: OpenApiResponse(
+                description="Unauthorized",
+                response={
+                    "type": "object",
+                    "properties": {
+                        "detail": {"type": "string", "example": "Authentication credentials were not provided."},
+                    },
+                },
+            ),
+        },
+    )
     def post(self, request):
-        request.user.auth_token.delete()
-        return Response({"message": "Successfully logged out."}, status=status.HTTP_200_OK)
+        if hasattr(request.user, 'auth_token'):
+            request.user.auth_token.delete()
+            return Response({"message": "Successfully logged out."}, status=status.HTTP_200_OK)
+        else:
+            return Response({"message": "No token found for this user."}, status=status.HTTP_400_BAD_REQUEST)
